Page 37 use scoped association to reduce calls
The example here is not exactly what the book intended
as I solved the problem with a join table on comments
and voters (users). The book wanted a `likes_count`
attribute on Comment, which I did add for the scoped
example. The unscoped example uses the join table.

I did not figure out how to preload the join table
for the unscoped example. This is opportunity I can
apply at work, and I may revisit.